Added short descriptions for a number of phases; recoloured colors for anglophiles.
While doing this, I changed many of the existing <short_desc>s to
phrases which were less descriptive, but more obviously related
to the abbreviated phase name. For example, I changed 

    jb.ule                       Removes unused locals

to

    jb.ule                       Unused local eliminator

on the rationale that it would then be easier for the reader of 
"soot --help" to understand where the name "jb.ule" came from.
This has the effect of making many <short_desc> values almost
identical to the corresponding <name> value.

git-svn-id: https://svn.sable.mcgill.ca/soot/soot/trunk@1017 8cfe803f-38b8-0310-ad8c-c0bb4caa2ed3