Changed the soot plugin Makefile.template so that it finds whatever version of a plugin is installed and compiles against that, rather than demanding a specific version.  Demanding a specific version is annoying when one has, say 2.1.1 of something and it wants 2.1.0.  Now setting the makefile variable V=_2.1* will find and use any sub-version of 2.1 (eg, either 2.1.0 or 2.1.1).
Also, my MacOSX installation of eclipse seems to have 2.1.0 of some packages
and 2.1.1 of other packages, so I didn't want to have to specify which version
I've got for each package.  The Linux and Windows installations of eclipse
seem to be more consistent: either all plugins are 2.1.0 or they're all
2.1.1 (at least all of the plugins that soot depends on).

Also ported the soot plugin Makefile.template to MacOSX.  Basically this just
meant adding the Carbon implementation of SWT to the classpath.

Tested on Redhat 9 and MacOSX 10.2.



git-svn-id: https://svn.sable.mcgill.ca/soot/soot/trunk@1152 8cfe803f-38b8-0310-ad8c-c0bb4caa2ed3