reworked strong must-alias analysis;
previously it could happen that an instance key was invalidated although it was not contained in a loop! this could happen if its definition was behind complicated control flow, i.e. the definition could be reached over more than one branch. now we use SCCs to determine whether a variable is redefined within a loop.

also, we moved the analysis and the class InstanceKey from abc to Soot

git-svn-id: https://svn.sable.mcgill.ca/soot/soot/trunk@3067 8cfe803f-38b8-0310-ad8c-c0bb4caa2ed3