Fixing problem described in posting "Class Hierarchy Analysis causes NullPointerException".
Comment by Ondrej:
When start() is invoked on a thread, a new thread is created and run()
is called implicitly on the newly-created thread. Thus the THREAD edge
is necessary to model the implicit call to run(). In this case, the
problem seems to be that start() is being called using a special invoke,
whereas the implicit call to run() is then (presumably) being done as
a virtual call. However, because the call to start() is a special
invoke, Soot tries to also resolve the call to run() as a special
invoke, which fails. The fix is to always handle the THREAD edge
as a virtual call, even if the original start() call is special.
So, the following if statement in addType:
   if( site.iie() instanceof SpecialInvokeExpr ) {
should be changed to:
   if( site.iie() instanceof SpecialInvokeExpr && site.kind != Kind.THREAD ) {

git-svn-id: https://svn.sable.mcgill.ca/soot/soot/trunk@3353 8cfe803f-38b8-0310-ad8c-c0bb4caa2ed3