Patch for phantom refs; work by Patrick Lam, Syed Albiz and Eric Bodden; still has some issues...
Firstly, after the patch, getPhantomRefs() and allowPhantomRefs() both return Options.v().allow_phantom_refs(). This means that, apparently, there is no need any more for the allowsPhantomRefs flag in the Scene, neither for the method setPhantomRefs(..). Correct?

Secondly, taking the above into account, tryLoadClass(..) seems to implement the exact same functionality as loadClass(..) after applying the patch. Can this be correct? Why were there two different methods to begin with?

Third, when creating a JInterfaceInvokeExpr and the target class is not an interface then after the patch, all we do is not report an exception if the target class is a phantom class. My intuition is that in this case we should also *make* the target class an interface in this case, basically inferring the fact that the class really should be an interface:
if(declaringClass.isPhantom()) {
declaringClass.setModifiers(declaringClass.getModifiers() | Modifier.INTERFACE);
}

git-svn-id: https://svn.sable.mcgill.ca/soot/soot/trunk@3359 8cfe803f-38b8-0310-ad8c-c0bb4caa2ed3